+++ src/main/java/org/nixos/idea/lang/highlighter/NixSyntaxHighlighterFactory.java
@@ -1,4 +1,4 @@
-package org.nixos.idea.lang;
+package org.nixos.idea.lang.highlighter;
 
 import com.intellij.openapi.fileTypes.SyntaxHighlighter;
 import com.intellij.openapi.fileTypes.SyntaxHighlighterFactory;
@@ -7,10 +7,8 @@ import com.intellij.openapi.vfs.VirtualFile;
 import org.jetbrains.annotations.NotNull;
 
 public class NixSyntaxHighlighterFactory extends SyntaxHighlighterFactory {
-    @NotNull
     @Override
-    public SyntaxHighlighter getSyntaxHighlighter(Project project, VirtualFile virtualFile) {
+    public @NotNull SyntaxHighlighter getSyntaxHighlighter(Project project, VirtualFile virtualFile) {
         return new NixSyntaxHighlighter();
     }
 }
-